Save iptables after flushing during cloud-init
The iptables rules should be saved after being flushed as part of the
cloud-init process. If the reboot plugin is used, the default CentOS
rules are loaded on boot because we are not saving the cleared set of
rules. This can cause some network configurations to break.
Ubuntu is not impacted as they do not have any 'default' rules.

Change-Id: I49842cd833a736f318dd237f07be2267dbfb27ab
Closes-Bug: 1459733