FA: oslo.serialization -> oslo_serialization
The issue is that oslo log module tries to
import oslo.serialization which is depricated.
Correct way to import this is to use
oslo_serialization module name.

The correct way how to fix this would be
installing oslo_log module instead of using
log module from oslo-incubator. But this way
is going to take long time because there are
no corresponding rpm and deb packages available.

TODO: use oslo_log instead of log module from oslo-incubator
when rpm and deb packages are available

Change-Id: I866fa0e19c5b872a82af6739d3537c24916bccc1
Closes-Bug: #1472018