Do not land /boot on NVMe disks
There's no point to place /boot partition on NVMe disk
as grub2 without UEFI support is unable to recognize it and
read any data from it.

Unfortunately, fuel-agent doesn't work with grub2-efi.
It always installs BIOS version of grub2.

It's just a work around that will allow to land /boot partition
on non-NVMe disk.

The absence of non-NVMe disks on a node will cause a failure of
provisioning.

DocImpact

Change-Id: I166c94416ccb152ccd8d1dc780dfb21a774a4f1d
Co-Authored-By: Atze de Vries <atze.devries@naturalis.nl>
Related-Bug: #1474970