Skip cleaning of MD containers
Add possibility to skip cleaning of MD containers which are usually
represent fake RAID devices. So, those devices won't be cleaned
and removed at the partitioning step. That will effectively save
those devices for further provisioning purposes.

In order to preserve existing behavior, it will be disabled by
default in conf file.

Will be enabled in a separate patch.

Change-Id: I7be1fc438ffe1fec9b8bfae7e47f50cd3a7c36b7
Related-Bug: #1508908