Change the way how we instantiate fuel client
The issue is that in python-fuelclient API has been
changed and fuelclient.client.APIClient is not available
anymore. It used to be a singleton of fuel client but
now it is deprecated. So, let's use APIClient.default_client
classmethod to instantiate APIClient.

https://github.com/openstack/python-fuelclient/blob/stable/newton/fuelclient/client.py#L51

Change-Id: I5287184c6f1c6abf8ac61894db0bd51fbf730401
Closes-Bug: #1620584
(cherry picked from commit 02b79798451a8d0680313a7ed9af390d114e0896)