Do not use separate partition for cloudinit configuration
In our usecases the separate partition is not needed. It is enough just
to put cloudinit configuration into the root filesystem.
This also allows to avoid a race condition which sometimes happens: some
process deletes the folder in tmp where the configuration partition is
mounted resulting in cloudinit failure to read its configuration.

Change-Id: Ib3efb4f517a5cf86dbf91ee53ac00108d4624dcd
Closes-Bug: #1652002